refactor(dtos): Dtos refactor

+++ src/main/java/com/chatop/chaTop/dto/MessageDto.java
@@ -1,4 +1,4 @@
 package com.chatop.chaTop.dto;
 
-public record MessageDto(String message, Long userId, Long rentalId) {
+public record MessageDto(String message, int userId, int rentalId) {
 }
