+++ src/main/java/com/chatop/chaTop/dto/UserDto.java
@@ -2,7 +2,7 @@ package com.chatop.chaTop.dto;
 
 import java.util.Date;
 
-public record UserDto(Long id,
+public record UserDto(int id,
                       String email, String name, Date createdAt, Date updatedAt) {
 
 }
